refactor and add some more tests
