+++ src/apphosting/utils.ts
@@ -1,3 +1,4 @@
+import { FirebaseError } from "../error";
 import { readFileFromDirectory, wrappedSafeLoad } from "../utils";
 import { Config as AppHostingYaml, AppHostingReadableConfiguration } from "./config";
 
@@ -27,4 +28,18 @@ export async function loadAppHostingYaml(
   }
 
   return { environmentVariables, secrets };
+}
+
+/**
+ * Returns <environment> given an apphosting.<environment>.yaml file
+ */
+export function getEnvironmentName(apphostingYamlFileName: string): string {
+  const envrionmentRegex = /apphosting\.(.+)\.yaml/;
+  const found = apphostingYamlFileName.match(envrionmentRegex);
+
+  if (!found) {
+    throw new FirebaseError("Invalid apphosting environment file");
+  }
+
+  return found[1];
 }
